+++ fl_api/lib/src/fl_api_header.freezed.dart
@@ -1,0 +1,362 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+
+part of 'fl_api_header.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+/// @nodoc
+class _$FlApiHeaderTearOff {
+  const _$FlApiHeaderTearOff();
+
+// ignore: unused_element
+  _FormData formData() {
+    return const _FormData();
+  }
+
+// ignore: unused_element
+  Basic basic() {
+    return const Basic();
+  }
+
+// ignore: unused_element
+  _Data data() {
+    return const _Data();
+  }
+}
+
+/// @nodoc
+// ignore: unused_element
+const $FlApiHeader = _$FlApiHeaderTearOff();
+
+/// @nodoc
+mixin _$FlApiHeader {
+  @optionalTypeArgs
+  Result when<Result extends Object>({
+    @required Result formData(),
+    @required Result basic(),
+    @required Result data(),
+  });
+  @optionalTypeArgs
+  Result maybeWhen<Result extends Object>({
+    Result formData(),
+    Result basic(),
+    Result data(),
+    @required Result orElse(),
+  });
+  @optionalTypeArgs
+  Result map<Result extends Object>({
+    @required Result formData(_FormData value),
+    @required Result basic(Basic value),
+    @required Result data(_Data value),
+  });
+  @optionalTypeArgs
+  Result maybeMap<Result extends Object>({
+    Result formData(_FormData value),
+    Result basic(Basic value),
+    Result data(_Data value),
+    @required Result orElse(),
+  });
+}
+
+/// @nodoc
+abstract class $FlApiHeaderCopyWith<$Res> {
+  factory $FlApiHeaderCopyWith(
+          FlApiHeader value, $Res Function(FlApiHeader) then) =
+      _$FlApiHeaderCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class _$FlApiHeaderCopyWithImpl<$Res> implements $FlApiHeaderCopyWith<$Res> {
+  _$FlApiHeaderCopyWithImpl(this._value, this._then);
+
+  final FlApiHeader _value;
+  // ignore: unused_field
+  final $Res Function(FlApiHeader) _then;
+}
+
+/// @nodoc
+abstract class _$FormDataCopyWith<$Res> {
+  factory _$FormDataCopyWith(_FormData value, $Res Function(_FormData) then) =
+      __$FormDataCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class __$FormDataCopyWithImpl<$Res> extends _$FlApiHeaderCopyWithImpl<$Res>
+    implements _$FormDataCopyWith<$Res> {
+  __$FormDataCopyWithImpl(_FormData _value, $Res Function(_FormData) _then)
+      : super(_value, (v) => _then(v as _FormData));
+
+  @override
+  _FormData get _value => super._value as _FormData;
+}
+
+/// @nodoc
+class _$_FormData implements _FormData {
+  const _$_FormData();
+
+  @override
+  String toString() {
+    return 'FlApiHeader.formData()';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) || (other is _FormData);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  @optionalTypeArgs
+  Result when<Result extends Object>({
+    @required Result formData(),
+    @required Result basic(),
+    @required Result data(),
+  }) {
+    assert(formData != null);
+    assert(basic != null);
+    assert(data != null);
+    return formData();
+  }
+
+  @override
+  @optionalTypeArgs
+  Result maybeWhen<Result extends Object>({
+    Result formData(),
+    Result basic(),
+    Result data(),
+    @required Result orElse(),
+  }) {
+    assert(orElse != null);
+    if (formData != null) {
+      return formData();
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  Result map<Result extends Object>({
+    @required Result formData(_FormData value),
+    @required Result basic(Basic value),
+    @required Result data(_Data value),
+  }) {
+    assert(formData != null);
+    assert(basic != null);
+    assert(data != null);
+    return formData(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  Result maybeMap<Result extends Object>({
+    Result formData(_FormData value),
+    Result basic(Basic value),
+    Result data(_Data value),
+    @required Result orElse(),
+  }) {
+    assert(orElse != null);
+    if (formData != null) {
+      return formData(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class _FormData implements FlApiHeader {
+  const factory _FormData() = _$_FormData;
+}
+
+/// @nodoc
+abstract class $BasicCopyWith<$Res> {
+  factory $BasicCopyWith(Basic value, $Res Function(Basic) then) =
+      _$BasicCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class _$BasicCopyWithImpl<$Res> extends _$FlApiHeaderCopyWithImpl<$Res>
+    implements $BasicCopyWith<$Res> {
+  _$BasicCopyWithImpl(Basic _value, $Res Function(Basic) _then)
+      : super(_value, (v) => _then(v as Basic));
+
+  @override
+  Basic get _value => super._value as Basic;
+}
+
+/// @nodoc
+class _$Basic implements Basic {
+  const _$Basic();
+
+  @override
+  String toString() {
+    return 'FlApiHeader.basic()';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) || (other is Basic);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  @optionalTypeArgs
+  Result when<Result extends Object>({
+    @required Result formData(),
+    @required Result basic(),
+    @required Result data(),
+  }) {
+    assert(formData != null);
+    assert(basic != null);
+    assert(data != null);
+    return basic();
+  }
+
+  @override
+  @optionalTypeArgs
+  Result maybeWhen<Result extends Object>({
+    Result formData(),
+    Result basic(),
+    Result data(),
+    @required Result orElse(),
+  }) {
+    assert(orElse != null);
+    if (basic != null) {
+      return basic();
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  Result map<Result extends Object>({
+    @required Result formData(_FormData value),
+    @required Result basic(Basic value),
+    @required Result data(_Data value),
+  }) {
+    assert(formData != null);
+    assert(basic != null);
+    assert(data != null);
+    return basic(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  Result maybeMap<Result extends Object>({
+    Result formData(_FormData value),
+    Result basic(Basic value),
+    Result data(_Data value),
+    @required Result orElse(),
+  }) {
+    assert(orElse != null);
+    if (basic != null) {
+      return basic(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class Basic implements FlApiHeader {
+  const factory Basic() = _$Basic;
+}
+
+/// @nodoc
+abstract class _$DataCopyWith<$Res> {
+  factory _$DataCopyWith(_Data value, $Res Function(_Data) then) =
+      __$DataCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class __$DataCopyWithImpl<$Res> extends _$FlApiHeaderCopyWithImpl<$Res>
+    implements _$DataCopyWith<$Res> {
+  __$DataCopyWithImpl(_Data _value, $Res Function(_Data) _then)
+      : super(_value, (v) => _then(v as _Data));
+
+  @override
+  _Data get _value => super._value as _Data;
+}
+
+/// @nodoc
+class _$_Data implements _Data {
+  const _$_Data();
+
+  @override
+  String toString() {
+    return 'FlApiHeader.data()';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) || (other is _Data);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  @optionalTypeArgs
+  Result when<Result extends Object>({
+    @required Result formData(),
+    @required Result basic(),
+    @required Result data(),
+  }) {
+    assert(formData != null);
+    assert(basic != null);
+    assert(data != null);
+    return data();
+  }
+
+  @override
+  @optionalTypeArgs
+  Result maybeWhen<Result extends Object>({
+    Result formData(),
+    Result basic(),
+    Result data(),
+    @required Result orElse(),
+  }) {
+    assert(orElse != null);
+    if (data != null) {
+      return data();
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  Result map<Result extends Object>({
+    @required Result formData(_FormData value),
+    @required Result basic(Basic value),
+    @required Result data(_Data value),
+  }) {
+    assert(formData != null);
+    assert(basic != null);
+    assert(data != null);
+    return data(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  Result maybeMap<Result extends Object>({
+    Result formData(_FormData value),
+    Result basic(Basic value),
+    Result data(_Data value),
+    @required Result orElse(),
+  }) {
+    assert(orElse != null);
+    if (data != null) {
+      return data(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class _Data implements FlApiHeader {
+  const factory _Data() = _$_Data;
+}